clinical-agent/requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
requests
groq
beautifulsoup4
